+++ domains/rem/src/hooks/useDatesBulkEditActions.ts
@@ -1,14 +1,17 @@
 import { useEffect } from 'react';
 
 import { __ } from '@eventespresso/i18n';
-import { hooks, Filters } from '@eventespresso/edtr-services';
+import { hooks, Actions, Filters } from '@eventespresso/edtr-services';
+import { useGlobalModal } from '@eventespresso/registry';
 
 import { NAMESPACE } from '../constants';
+import { RemGlobalModals } from '../types';
 
 const filterName: keyof Filters = 'eventEditor.datetimes.bulkEdit.actions';
+const actionName: keyof Actions = 'eventEditor.datetimes.bulkEdit.apply';
 
 /**
- * A custom hook to to add WP User section to Ticket edit form
+ * A custom hook to to handle bulk add tickets to dates
  */
 const useDatesBulkEditActions = (): void => {
 	useEffect(() => {
@@ -24,6 +27,21 @@ const useDatesBulkEditActions = (): void => {
 
 		// housekeeping
 		return () => hooks.removeFilter(filterName, NAMESPACE);
+	}, []);
+
+	const { openWithData } = useGlobalModal(RemGlobalModals.BULK_ADD_TICKETS);
+	useEffect(() => {
+		hooks.addAction(actionName, NAMESPACE, (action, bulkEdit) => {
+			if (action !== 'bulk-add-tickets') {
+				return;
+			}
+			const entityIds = bulkEdit.getSelected();
+			openWithData({ entityIds });
+		});
+
+		// housekeeping
+		return () => hooks.removeAction(actionName, NAMESPACE);
+		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, []);
 };
